backend/requirements.txt
flask
flask-cors
pandas
google-generativeai
google-auth
requests
captcha
google-cloud-speech
gtts
playsound
SpeechRecognition